+++ Unifiedban.Terminal/Bot/Command/GetTrustFactor.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Linq;
+using Telegram.Bot;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Unifiedban.Terminal.Utils;
@@ -14,7 +15,7 @@ namespace Unifiedban.Terminal.Bot.Command
     {
         public void Execute(Message message)
         {
-            int userId;
+            long userId;
 
             if (message.ReplyToMessage == null)
             {
@@ -39,7 +40,7 @@ namespace Unifiedban.Terminal.Bot.Command
                 }
                 else
                 {
-                    bool isValid = int.TryParse(message.Text.Split(" ")[1], out userId);
+                    bool isValid = long.TryParse(message.Text.Split(" ")[1], out userId);
                     if (!isValid)
                     {
                         MessageQueueManager.EnqueueMessage(
