+++ Unifiedban.Terminal/Bot/Command/Kick.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Linq;
+using Telegram.Bot;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Unifiedban.Terminal.Utils;
@@ -29,7 +30,7 @@ namespace Unifiedban.Terminal.Bot.Command
                    });
                 return;
             }
-            int userToKick;
+            long userToKick;
 
             if (message.ReplyToMessage == null)
             {
@@ -50,7 +51,7 @@ namespace Unifiedban.Terminal.Bot.Command
                 }
                 else
                 {
-                    bool isValid = int.TryParse(message.Text.Split(" ")[1], out userToKick);
+                    bool isValid = long.TryParse(message.Text.Split(" ")[1], out userToKick);
                     if (!isValid)
                     {
                         MessageQueueManager.EnqueueMessage(
