+++ Unifiedban.Terminal/Bot/Command/Unmute.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Telegram.Bot;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 
@@ -24,14 +25,14 @@ namespace Unifiedban.Terminal.Bot.Command
                 return;
             }
 
-            int userId = 0;
+            long userId = 0;
             if (message.ReplyToMessage != null)
             {
                 userId = message.ReplyToMessage.From.Id;
             }
             else if (message.Text.Contains(" "))
             {
-                int.TryParse(message.Text.Split(" ")[1], out userId);
+                long.TryParse(message.Text.Split(" ")[1], out userId);
             }
 
             if (userId == 0)
